[refactor] Adjust file location and docs
