+++ transforms/__init__.py
@@ -1,5 +1,8 @@
 from .builder import TRANSFORMS, build_transform
 from .transform import *  # noqa: F401,F403
+from .plugins.misc import *  # noqa: F401,F403
+from .plugins.dextr import *  # noqa: F401,F403
+from .plugins.albumentations import *  # noqa: F401,F403
 
 __all__ = [
     TRANSFORMS,
